refactor ObserverUtils
