+++ dspot/src/main/java/eu/stamp_project/dspot/assertgenerator/components/TestMethodReconstructor.java
@@ -4,7 +4,7 @@ import eu.stamp_project.compare.Observation;
 import eu.stamp_project.dspot.AmplificationException;
 import eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor.AssertionSyntaxBuilder;
 import eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor.Observer;
-import eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor.observer.ObserverUtils;
+import eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor.observer.TestWithLogGenerator;
 import eu.stamp_project.dspot.assertgenerator.components.utils.AssertionGeneratorUtils;
 import eu.stamp_project.utils.AmplificationHelper;
 import eu.stamp_project.utils.CloneHelper;
@@ -70,7 +70,7 @@ public class TestMethodReconstructor {
      * Adds new assertions in multiple tests.
      * <p>
      * <p>Instruments the tests to have observation points.
-     * Details in {@link ObserverUtils#createTestWithLog(CtMethod, String, List)}.
+     * Details in {@link TestWithLogGenerator#createTestWithLog(CtMethod, String, List)}.
      * <p>
      * <p>Details of the assertion generation in {@link #buildTestWithAssert(CtMethod, Map)}.
      *
