+++ dspot/src/main/java/eu/stamp_project/dspot/assertgenerator/components/testmethodreconstructor/observer/components/LogBuilder.java
@@ -1,129 +1,25 @@
-package eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor.observer;
+package eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor.observer.components;
 
-import eu.stamp_project.compare.MethodsHandler;
 import eu.stamp_project.compare.ObjectLog;
 import eu.stamp_project.dspot.assertgenerator.components.utils.AssertionGeneratorUtils;
-import eu.stamp_project.utils.CloneHelper;
 import spoon.reflect.code.*;
 import spoon.reflect.declaration.CtMethod;
-import spoon.reflect.declaration.CtNamedElement;
-import spoon.reflect.declaration.CtTypedElement;
 import spoon.reflect.factory.Factory;
 import spoon.reflect.reference.CtExecutableReference;
-import spoon.reflect.reference.CtTypeReference;
-import spoon.reflect.reference.CtWildcardReference;
-import spoon.reflect.visitor.filter.TypeFilter;
-import spoon.support.SpoonClassNotFoundException;
 import java.util.List;
 import java.util.function.Predicate;
 
-public class ObserverUtils {
+public class LogBuilder {
 
     private static final Predicate<CtStatement> shouldAddLogEndStatement = shouldAddLogEndStatement();
 
     /**
-     *
-     * @param test
-     * @param filter
-     * @param ctVariableReads
-     * @return
+     * This method will add a log statement at the given statement AND at the end of the test.
+     * @param stmt
+     * @param id
      */
-    public static CtMethod<?> createTestWithLog(CtMethod test, final String filter,
-                                                List<CtLocalVariable<?>> ctVariableReads) {
-        CtMethod clone = CloneHelper.cloneTestMethodNoAmp(test);
-        clone.setSimpleName(test.getSimpleName() + "_withlog");
-        final List<CtStatement> allStatement = clone.getElements(new TypeFilter<>(CtStatement.class));
-        allStatement.stream()
-                .filter(statement ->
-                        (ObserverUtils.isStmtToLog(filter, statement) ||
-                                ctVariableReads.contains(statement)) &&
-                                isNotFromPreviousAmplification(allStatement, statement, test)
-                ).forEach(statement ->
-                addLogStmt(statement,
-                        test.getSimpleName() + "__" + indexOfByRef(allStatement, statement))
-        );
-        return clone;
-    }
-
-    private static boolean isStmtToLog(String filter, CtStatement statement) {
-        if (!(statement.getParent() instanceof CtBlock)) {
-            return false;
-        }
-
-        // contract: for now, we do not log values inside loop
-        if (statement.getParent(CtLoop.class) != null) {
-            return false;
-        }
-        if (statement instanceof CtInvocation) {
-            CtInvocation invocation = (CtInvocation) statement;
-            return (invocation.getMetadata(AssertionGeneratorUtils.METADATA_WAS_IN_ASSERTION) != null &&
-                    (Boolean) invocation.getMetadata(AssertionGeneratorUtils.METADATA_WAS_IN_ASSERTION)) ||
-                    (isCorrectReturn(invocation) && !isGetter(invocation));
-        }
-        if (statement instanceof CtLocalVariable ||
-                statement instanceof CtAssignment ||
-                statement instanceof CtVariableWrite) {
-            if (statement instanceof CtNamedElement) {
-                if (((CtNamedElement) statement).getSimpleName()
-                        .startsWith("__DSPOT_")) {
-                    return false;
-                }
-            }
-            final CtTypeReference type = ((CtTypedElement) statement).getType();
-            if (type.getQualifiedName().startsWith(filter)) {
-                return true;
-            } else {
-                try {
-                    return type.getTypeDeclaration().getQualifiedName()
-                            .equals("java.lang.String");
-                } catch (SpoonClassNotFoundException e) {
-                    return false;
-                }
-            }
-        } else {
-            return false;
-        }
-    }
-
-    private static boolean isGetter(CtInvocation invocation) {
-        return invocation.getArguments().isEmpty() &&
-                MethodsHandler.isASupportedMethodName(invocation.getExecutable().getSimpleName());
-    }
-
-    private static boolean isCorrectReturn(CtInvocation<?> invocation) {
-        return invocation.getType() != null &&
-                !(AssertionGeneratorUtils.isVoidReturn(invocation)) &&
-                !(invocation.getType() instanceof CtWildcardReference) &&
-                invocation.getType().getTypeDeclaration() != null &&
-                !("java.lang.Class".equals(invocation.getType().getTypeDeclaration().getQualifiedName()));
-    }
-
-    // This method aims at infer from the name of the local variables if it came from a previous amplification.
-    // see the use case available here: https://github.com/STAMP-project/dspot/issues/825
-    private static boolean isNotFromPreviousAmplification(final List<CtStatement> allStatement,
-                                                          CtStatement statement,
-                                                          CtMethod<?> test) {
-        final String id = test.getSimpleName() + "__" + ObserverUtils.indexOfByRef(allStatement, statement);
-        return test.getElements(new TypeFilter<CtLocalVariable>(CtLocalVariable.class) {
-            @Override
-            public boolean matches(CtLocalVariable element) {
-                return element.getSimpleName().endsWith(id);
-            }
-        }).isEmpty();
-    }
-
-    private static int indexOfByRef(List<CtStatement> statements, CtStatement statement) {
-        for (int i = 0; i < statements.size(); i++) {
-            if (statements.get(i) == statement) {
-                return i;
-            }
-        }
-        return -1;
-    }
-
-    // This method will add a log statement at the given statement AND at the end of the test.
     @SuppressWarnings("unchecked")
-    private static void addLogStmt(CtStatement stmt, String id) {
+    public static void addLogStmt(CtStatement stmt, String id) {
         if (stmt instanceof CtLocalVariable && ((CtLocalVariable) stmt).getDefaultExpression() == null) {
             return;
         }
@@ -159,11 +55,12 @@ public class ObserverUtils {
         CtInvocation invocationToObjectLogAtTheEnd = insertStatement(stmt,factory,invocationToObjectLog,id,insertAfter);
 
         // if between the two log statements there is only log statement, we do not add the log end statement
-        if (ObserverUtils.shouldAddLogEndStatement.test(invocationToObjectLog) &&
-                ObserverUtils.getSize(stmt.getParent(CtMethod.class).getBody()) + 1 < 65535) {
+        if (shouldAddLogEndStatement.test(invocationToObjectLog) &&
+                getSize(stmt.getParent(CtMethod.class).getBody()) + 1 < 65535) {
             stmt.getParent(CtBlock.class).insertEnd(invocationToObjectLogAtTheEnd);
         }
     }
+
 
     private static CtStatement localVariableLogStmt(CtStatement stmt, Factory factory, CtInvocation invocationToObjectLog){
         CtLocalVariable localVar = (CtLocalVariable) stmt;
@@ -236,13 +133,13 @@ public class ObserverUtils {
 
     // clone the statement invocation for add it to the end of the tests
     private static CtInvocation insertStatement(CtStatement stmt, Factory factory,
-                                 CtInvocation invocationToObjectLog, String id, CtStatement insertAfter){
+                                                CtInvocation invocationToObjectLog, String id, CtStatement insertAfter){
         CtInvocation invocationToObjectLogAtTheEnd = invocationToObjectLog.clone();
         invocationToObjectLogAtTheEnd.addArgument(factory.createLiteral(id + "___" + "end"));
         invocationToObjectLog.addArgument(factory.createLiteral(id));
 
         //TODO checks this if this condition is ok.
-        if (ObserverUtils.getSize(stmt.getParent(CtMethod.class).getBody()) + 1 < 65535) {
+        if (getSize(stmt.getParent(CtMethod.class).getBody()) + 1 < 65535) {
             insertAfter.insertAfter(invocationToObjectLog);
         }
         return invocationToObjectLogAtTheEnd;
