+++ dspot/src/test/java/eu/stamp_project/dspot/assertgenerator/AssertionGeneratorUtilsTest.java
@@ -4,7 +4,7 @@ import eu.stamp_project.AbstractTest;
 import eu.stamp_project.Utils;
 import eu.stamp_project.dspot.amplifier.MethodGeneratorAmplifier;
 import eu.stamp_project.dspot.assertgenerator.components.AssertionRemover;
-import eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor.observer.ObserverUtils;
+import eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor.observer.TestWithLogGenerator;
 import eu.stamp_project.dspot.assertgenerator.components.utils.AssertionGeneratorUtils;
 import eu.stamp_project.test_framework.TestFramework;
 import eu.stamp_project.utils.AmplificationHelper;
@@ -193,7 +193,7 @@ public class AssertionGeneratorUtilsTest extends AbstractTest {
                         "    eu.stamp_project.compare.ObjectLog.log(o_test2__5, \"o_test2__5\", \"test2__5___end\");" + AmplificationHelper.LINE_SEPARATOR +
                         "    eu.stamp_project.compare.ObjectLog.log(o_test2__7, \"o_test2__7\", \"test2__7___end\");" + AmplificationHelper.LINE_SEPARATOR +
                         "}",
-                ObserverUtils.createTestWithLog(
+                TestWithLogGenerator.createTestWithLog(
                         new AssertionRemover().removeAssertion(Utils.findMethod("fr.inria.sample.TestClassWithLoop", "test2")),
                         "fr.inria.sample",
                         Collections.emptyList()).toString()
@@ -222,7 +222,7 @@ public class AssertionGeneratorUtilsTest extends AbstractTest {
         assertEquals(4, amplifiedMethods.size());
 
         final List<CtMethod<?>> instrumentedAmplifiedTests = amplifiedMethods.stream()
-                .map(method -> ObserverUtils.createTestWithLog(method, "fr.inria.statementaddarray", Collections.emptyList()))
+                .map(method -> TestWithLogGenerator.createTestWithLog(method, "fr.inria.statementaddarray", Collections.emptyList()))
                 .collect(Collectors.toList());
 
         assertEquals(4, instrumentedAmplifiedTests.size());
@@ -245,7 +245,7 @@ public class AssertionGeneratorUtilsTest extends AbstractTest {
     public void testMultipleObservationsPoints() throws Exception {
         final CtMethod<?> test1 = Utils.findMethod("fr.inria.multipleobservations.TestClassToBeTest", "test");
         final CtMethod<?> testWithLog =
-                ObserverUtils.createTestWithLog(test1, "fr.inria.multipleobservations", Collections.emptyList());
+                TestWithLogGenerator.createTestWithLog(test1, "fr.inria.multipleobservations", Collections.emptyList());
         final String expectedMethodWithLogs = "@org.junit.Test(timeout = 10000)" + AmplificationHelper.LINE_SEPARATOR +
                 "public void test_withlog() throws java.lang.Exception {" + AmplificationHelper.LINE_SEPARATOR +
                 "    final fr.inria.multipleobservations.ClassToBeTest classToBeTest = new fr.inria.multipleobservations.ClassToBeTest();" + AmplificationHelper.LINE_SEPARATOR +
@@ -260,7 +260,7 @@ public class AssertionGeneratorUtilsTest extends AbstractTest {
     public void testCreateTestWithLogClassTargetAmplify() throws Exception {
         final CtMethod<?> test1 = Utils.findMethod("fr.inria.statementaddarray.TestClassTargetAmplify", "test");
         final CtMethod<?> testWithLog =
-                ObserverUtils.createTestWithLog(test1, "fr.inria.statementaddarray", Collections.emptyList());
+                TestWithLogGenerator.createTestWithLog(test1, "fr.inria.statementaddarray", Collections.emptyList());
         final String expectedMethod = "@org.junit.Test(timeout = 10000)" + AmplificationHelper.LINE_SEPARATOR +
                 "public void test_withlog() throws java.lang.Exception {" + AmplificationHelper.LINE_SEPARATOR +
                 "    fr.inria.statementaddarray.ClassTargetAmplify clazz = new fr.inria.statementaddarray.ClassTargetAmplify();" + AmplificationHelper.LINE_SEPARATOR +
@@ -281,7 +281,7 @@ public class AssertionGeneratorUtilsTest extends AbstractTest {
         CtClass testClass = Utils.findClass("fr.inria.sample.TestClassWithoutAssert");
         final CtMethod<?> test1 = (CtMethod<?>) testClass.getMethodsByName("test1").get(0);
         final CtMethod<?> testWithLog =
-                ObserverUtils.createTestWithLog(test1, "fr.inria.sample", Collections.emptyList());
+                TestWithLogGenerator.createTestWithLog(test1, "fr.inria.sample", Collections.emptyList());
 
         final String expectedMethod = "@org.junit.Test(timeout = 10000)" + AmplificationHelper.LINE_SEPARATOR +
                 "public void test1_withlog() throws java.lang.Exception {" + AmplificationHelper.LINE_SEPARATOR +
@@ -300,7 +300,7 @@ public class AssertionGeneratorUtilsTest extends AbstractTest {
     public void testCreateTestWithLogWithoutChainSameObservations() throws Exception {
         CtMethod test1 = Utils.findMethod("fr.inria.sample.TestClassWithSpecificCaseToBeAsserted", "test1");
         final CtMethod<?> testWithLog =
-                ObserverUtils.createTestWithLog(test1, "fr.inria.sample", Collections.emptyList());
+                TestWithLogGenerator.createTestWithLog(test1, "fr.inria.sample", Collections.emptyList());
 
         final String expectedMethodWithLog = "@org.junit.Test(timeout = 10000)" + AmplificationHelper.LINE_SEPARATOR +
                 "public void test1_withlog() throws java.lang.Exception {" + AmplificationHelper.LINE_SEPARATOR +
@@ -325,7 +325,7 @@ public class AssertionGeneratorUtilsTest extends AbstractTest {
         CtClass testClass = Utils.findClass("fr.inria.sample.TestClassWithoutAssert");
         final CtMethod<?> test2 = (CtMethod<?>) testClass.getMethodsByName("test2").get(0);
         final CtMethod<?> testWithLog =
-                ObserverUtils.createTestWithLog(test2, "fr.inria.sample", Collections.emptyList());
+                TestWithLogGenerator.createTestWithLog(test2, "fr.inria.sample", Collections.emptyList());
 
         final String expectedMethod = "@org.junit.Test(timeout = 10000)" + AmplificationHelper.LINE_SEPARATOR +
                 "public void test2_withlog() throws java.lang.Exception {" + AmplificationHelper.LINE_SEPARATOR +
